requirements.txt
openai
langchain
python-dotenv
huggingface_hub
langchain-community
tensorflow
matplotlib
numpy
pandas